requirements_fastmri.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
absl-py=2.3.1=pypi_0
addict=2.4.0=pypi_0
basicsr=1.4.2=pypi_0
bzip2=1.0.8=h2bbff1b_6
ca-certificates=2025.2.25=haa95532_0
certifi=2025.7.14=pypi_0
charset-normalizer=3.4.2=pypi_0
colorama=0.4.6=pypi_0
contourpy=1.3.0=pypi_0
cycler=0.12.1=pypi_0
expat=2.7.1=h8ddb27b_0
facexlib=0.3.0=pypi_0
filelock=3.18.0=pypi_0
filterpy=1.4.5=pypi_0
fonttools=4.59.0=pypi_0
fsspec=2025.7.0=pypi_0
future=1.0.0=pypi_0
gfpgan=1.3.8=pypi_0
grpcio=1.74.0=pypi_0
h5py=3.14.0=pypi_0
idna=3.10=pypi_0
imageio=2.37.0=pypi_0
importlib-metadata=8.7.0=pypi_0
importlib-resources=6.5.2=pypi_0
jinja2=3.1.6=pypi_0
kiwisolver=1.4.7=pypi_0
lazy-loader=0.4=pypi_0
libffi=3.4.4=hd77b12b_1
llvmlite=0.43.0=pypi_0
lmdb=1.7.3=pypi_0
markdown=3.8.2=pypi_0
markupsafe=3.0.2=pypi_0
matplotlib=3.9.4=pypi_0
mpmath=1.3.0=pypi_0
networkx=3.2.1=pypi_0
numba=0.60.0=pypi_0
numpy=2.0.2=pypi_0
opencv-python=4.12.0.88=pypi_0
openssl=3.0.17=h35632f6_0
packaging=25.0=pypi_0
pillow=11.3.0=pypi_0
pip=25.1=pyhc872135_2
platformdirs=4.3.8=pypi_0
protobuf=6.31.1=pypi_0
pyparsing=3.2.3=pypi_0
python=3.9.23=h716150d_0
python-dateutil=2.9.0.post0=pypi_0
pyyaml=6.0.2=pypi_0
realesrgan=0.3.0=dev_0
requests=2.32.4=pypi_0
scikit-image=0.24.0=pypi_0
scipy=1.13.1=pypi_0
setuptools=78.1.1=py39haa95532_0
six=1.17.0=pypi_0
sqlite=3.50.2=hda9a48d_1
sympy=1.13.1=pypi_0
tb-nightly=2.21.0a20250725=pypi_0
tensorboard-data-server=0.7.2=pypi_0
tifffile=2024.8.30=pypi_0
tk=8.6.14=h5e9d12e_1
tomli=2.2.1=pypi_0
torch=2.5.1+cu121=pypi_0
torchaudio=2.5.1+cu121=pypi_0
torchvision=0.20.1+cu121=pypi_0
tqdm=4.67.1=pypi_0
typing-extensions=4.14.1=pypi_0
tzdata=2025b=h04d1e81_0
ucrt=10.0.22621.0=haa95532_0
urllib3=2.5.0=pypi_0
vc=14.3=h2df5915_10
vc14_runtime=14.44.35208=h4927774_10
vs2015_runtime=14.44.35208=ha6b5a95_10
werkzeug=3.1.3=pypi_0
wheel=0.45.1=py39haa95532_0
xz=5.6.4=h4754444_1
yapf=0.43.0=pypi_0
zipp=3.23.0=pypi_0
zlib=1.2.13=h8cc25b3_1
